Начало 3-го спринта, перенес mainpage в router
