+++ react-burger/package.json
@@ -19,7 +19,7 @@
     "react-dnd-html5-backend": "^16.0.1",
     "react-dom": "^18.2.0",
     "react-redux": "^8.1.2",
-    "react-router-dom": "^6.14.2",
+    "react-router-dom": "^6.15.0",
     "react-scripts": "5.0.1",
     "react-uuid": "^2.0.0",
     "redux": "^4.2.1",
