+++ react-burger/src/pages/page-404.jsx
@@ -1,0 +1,3 @@
+export const Page404 = () => {
+    
+}
